requirements.txt
torchdata>=0.8.0
spacy>=3.7.6
python-dotenv>=1.0.1
tqdm>=4.64.1
portalocker>=2.10.1
huggingface-hub>=0.10.1